requirements.txt
google-adk
streamlit
fastapi
pydantic
pandas
python-dotenv
uvicorn
